+++ signalwire/action_prompt.go
@@ -520,3 +520,14 @@ func (action *PromptAction) GetPayload() *json.RawMessage {
 
 	return ret
 }
+
+// GetControlID TODO DESCRIPTION
+func (action *PromptAction) GetControlID() string {
+	action.RLock()
+
+	ret := action.ControlID
+
+	action.RUnlock()
+
+	return ret
+}